Replace xmlpath with xquery
Replaces `xmlpath` library with the `xquery` library.

The `xmlpath` library, while not broken, is licensed under an LGPL license, and therefor is unable to be used by our organization.

This replaces the `xmlpath` library with a fully functional `xquery` library that is licensed under the MIT license.

This serves as a direct replacement without incurring any breaking changes to the libraries API, or any un-warranted side-effects.

```
$ go test -v
=== RUN   Test_creating_a_shell
--- PASS: Test_creating_a_shell (0.00s)
=== RUN   Test_executing_a_command
--- PASS: Test_executing_a_command (0.00s)
=== RUN   Test_executing_a_regex_command
--- PASS: Test_executing_a_regex_command (0.00s)
=== RUN   Test_receiving_command_results
--- PASS: Test_receiving_command_results (0.00s)
=== RUN   Test_deleting_a_shell
--- PASS: Test_deleting_a_shell (0.00s)
PASS
ok      github.com/dylanmei/winrmtest   0.003s
```